+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Updated Exceptional call signature
+* Added @phpstan-require-implements constraints
+
 ## v0.4.7 (2024-08-21)
 * Converted consts to protected PascalCase
 * Updated Veneer dependency and Stub